+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/bridge/influxdb/handler/InfluxdbBridgeNormalDataHandler.java
@@ -1,0 +1,10 @@
+package com.dwarfeng.fdr.impl.handler.bridge.influxdb.handler;
+
+/**
+ * Influxdb 桥接器一般数据处理器。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
+public interface InfluxdbBridgeNormalDataHandler extends InfluxdbBridgeDataHandler {
+}