+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/bridge/influxdb/util/Constants.java
@@ -1,0 +1,30 @@
+package com.dwarfeng.fdr.impl.handler.bridge.influxdb.util;
+
+/**
+ * 常量类。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
+public final class Constants {
+
+    /**
+     * InfluxDB 字段名：数据值。
+     */
+    public static final String FIELD_NAME_VALUE = "value";
+
+    /**
+     * InfluxDB 字段名：过滤器 ID。
+     */
+    public static final String FIELD_NAME_FILTER_ID = "filter_id";
+
+    /**
+     * InfluxDB 字段名：触发器 ID。
+     */
+    public static final String FIELD_NAME_TRIGGER_ID = "trigger_id";
+
+    /**
+     * InfluxDB 字段名：消息。
+     */
+    public static final String FILED_NAME_MESSAGE = "message";
+}