+++ fdr-node/src/main/java/com/dwarfeng/fdr/node/configuration/JulToSlf4jLogConfiguration.java
@@ -1,0 +1,19 @@
+package com.dwarfeng.fdr.node.configuration;
+
+import org.slf4j.bridge.SLF4JBridgeHandler;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+
+import javax.annotation.PostConstruct;
+
+@Configuration
+@Order(Ordered.HIGHEST_PRECEDENCE)
+public class JulToSlf4jLogConfiguration {
+
+    @PostConstruct
+    public void init() {
+        SLF4JBridgeHandler.removeHandlersForRootLogger();
+        SLF4JBridgeHandler.install();
+    }
+}